add support for remote config
The remote config facility makes it possible for clients to override the
default server side config by passing a config block in the first
message in the session.

The config message has fields for IP address and port and makes it
possible for the client to select which server the server side instance
should connect to.

Signed-off-by: Erik Stromdahl <erik.stromdahl@gmail.com>